Updated files with additional changes after initial PR
