+++ main.rb
@@ -1,0 +1,97 @@
+require_relative 'country'
+require_relative 'loan_eligibility'
+require_relative 'un_security_council'
+require_relative 'war_prediction'
+
+$all_countries = []
+
+def find_country(name)
+  $all_countries.find { |c| c.name.downcase == name.downcase }
+end
+
+def menu
+  loop do
+    puts "\n=== Country Intelligence CLI ==="
+    puts "1. Add new country"
+    puts "2. Check loan eligibility"
+    puts "3. Check UN Security Council eligibility"
+    puts "4. Predict war outcome"
+    puts "5. List all countries"
+    puts "6. Exit"
+    print "Choose an option: "
+    choice = gets.chomp.to_i
+
+    case choice
+    when 1
+      add_country
+    when 2
+      check_loan
+    when 3
+      check_un
+    when 4
+      predict_war
+    when 5
+      list_countries
+    when 6
+      puts "Exiting!!"
+      break
+    else
+      puts "Invalid choice!!"
+    end
+  end
+end
+
+def add_country
+  print "Enter country name: "
+  name = gets.chomp
+  print "Enter population: "
+  population = gets.chomp.to_i
+  print "Enter GDP (in millions USD): "
+  gdp = gets.chomp.to_i
+  print "Enter number of states: "
+  states = gets.chomp.to_i
+  print "Enter army strength: "
+  army = gets.chomp.to_i
+  print "Enter development status (developed/developing): "
+  status = gets.chomp.downcase
+  print "Enter region: "
+  region = gets.chomp
+
+  country = Country.new(name, population, gdp, states, army, status, region)
+  $all_countries << country
+  puts "\nCountry added: #{country.name}"
+end
+
+def check_loan
+  print "Enter country name: "
+  name = gets.chomp
+  country = find_country(name)
+  puts country ? country.loan_eligibility? : "Country not found."
+end
+
+def predict_war
+  print "Enter country name: "
+  name = gets.chomp
+  country = find_country(name)
+  puts country ? country.war_outcome : "Country not found."
+end
+
+def check_un
+  print "Enter country name: "
+  name = gets.chomp
+  country = find_country(name)
+  puts country ? country.un_security_council_eligibility? : "Country not found."
+end
+
+def list_countries
+  if $all_countries.empty?
+    puts "No countries in the system yet."
+  else
+    puts "\nList of countries:"
+    $all_countries.each_with_index do |c, i|
+      puts "#{i + 1}. #{c.name} (#{c.development_status.capitalize}, GDP: #{c.gdp}M)"
+    end
+  end
+end
+
+menu
